refactor: creation of hexagonal folders structure
