+++ src/main/scala/entities/domain/train/Wagons.scala
@@ -1,4 +1,4 @@
-package train.model.domain
+package entities.domain.train
 
 object Wagons:
   private type Capacity = Int
